# Optimize latin languages detection