--- a/Cargo.toml
+++ b/Cargo.toml
@@ -23,6 +23,7 @@ github-actions = { repository = "greyblake/whatlang-rs", workflow = "CI", branch
 
 [dependencies]
 hashbrown = "0.7"
+once_cell = "1.10.0"
 enum-map = { version = "0.6", optional = true }
 
 [dev-dependencies]