--- a/misc/alphabets/calculate_scores.rs.erb
+++ b/misc/alphabets/calculate_scores.rs.erb
@@ -43,7 +43,7 @@ pub fn alphabet_calculate_scores(text: &str) -> Outcome {
         }
     }
 
-    raw_scores.sort_by(|a, b| b.1.cmp(&a.1));
+    raw_scores.sort_unstable_by(|a, b| b.1.cmp(&a.1));
 
     let raw_scores: Vec<(Lang, usize)> = raw_scores
         .into_iter()