--- a/src/alphabets/cyrillic.rs
+++ b/src/alphabets/cyrillic.rs
@@ -1,3 +1,5 @@
+use std::cmp::Reverse;
+
 use super::RawOutcome;
 use crate::core::{FilterList, LowercaseText};
 use crate::{Lang, Script};
@@ -35,7 +37,7 @@ pub fn alphabet_calculate_scores(text: &LowercaseText, filter_list: &FilterList)
         }
     }
 
-    raw_scores.sort_unstable_by(|a, b| b.1.cmp(&a.1));
+    raw_scores.sort_unstable_by_key(|(_, score)| Reverse(*score));
 
     let raw_scores: Vec<(Lang, usize)> = raw_scores
         .into_iter()