--- a/src/alphabets/latin.rs
+++ b/src/alphabets/latin.rs
@@ -3,6 +3,9 @@ use crate::core::{FilterList, LowercaseText};
 use crate::utils::is_stop_char;
 use crate::{Lang, Script};
 
+use once_cell::sync::Lazy;
+use std::collections::HashMap;
+
 const AFR: &str = "abcdefghijklmnopqrstuvwxyzáèéêëíîïóôúû";
 const AKA: &str = "abdefghiklmnoprstuwyɔɛ";
 const AZE: &str = "abcdefghijklmnopqrstuvxyzçöüğışə̇";
@@ -41,71 +44,86 @@ const VIE: &str =
     "abcdefghijklmnopqrstuvwxyzàáâãèéêìíòóôõùúýăđĩũơưạảấầẩẫậắằẳẵặẹẻẽếềểễệỉịọỏốồổỗộớờởỡợụủứừửữựỳỵỷỹ";
 const ZUL: &str = "abcdefghijklmnopqrstuvwxyz";
 
-fn get_lang_chars(lang: Lang) -> Vec<char> {
-    let alphabet = match lang {
-        Lang::Afr => AFR,
-        Lang::Aka => AKA,
-        Lang::Aze => AZE,
-        Lang::Cat => CAT,
-        Lang::Ces => CES,
-        Lang::Dan => DAN,
-        Lang::Deu => DEU,
-        Lang::Eng => ENG,
-        Lang::Epo => EPO,
-        Lang::Est => EST,
-        Lang::Fin => FIN,
-        Lang::Fra => FRA,
-        Lang::Hrv => HRV,
-        Lang::Hun => HUN,
-        Lang::Ind => IND,
-        Lang::Ita => ITA,
-        Lang::Jav => JAV,
-        Lang::Lat => LAT,
-        Lang::Lav => LAV,
-        Lang::Lit => LIT,
-        Lang::Nld => NLD,
-        Lang::Nob => NOB,
-        Lang::Pol => POL,
-        Lang::Por => POR,
-        Lang::Ron => RON,
-        Lang::Slk => SLK,
-        Lang::Slv => SLV,
-        Lang::Sna => SNA,
-        Lang::Spa => SPA,
-        Lang::Swe => SWE,
-        Lang::Tgl => TGL,
-        Lang::Tuk => TUK,
-        Lang::Tur => TUR,
-        Lang::Uzb => UZB,
-        Lang::Vie => VIE,
-        Lang::Zul => ZUL,
+const LATIN_ALPHABETS: &[(Lang, &str)] = &[
+    (Lang::Afr, AFR),
+    (Lang::Aka, AKA),
+    (Lang::Aze, AZE),
+    (Lang::Cat, CAT),
+    (Lang::Ces, CES),
+    (Lang::Dan, DAN),
+    (Lang::Deu, DEU),
+    (Lang::Eng, ENG),
+    (Lang::Epo, EPO),
+    (Lang::Est, EST),
+    (Lang::Fin, FIN),
+    (Lang::Fra, FRA),
+    (Lang::Hrv, HRV),
+    (Lang::Hun, HUN),
+    (Lang::Ind, IND),
+    (Lang::Ita, ITA),
+    (Lang::Jav, JAV),
+    (Lang::Lat, LAT),
+    (Lang::Lav, LAV),
+    (Lang::Lit, LIT),
+    (Lang::Nld, NLD),
+    (Lang::Nob, NOB),
+    (Lang::Pol, POL),
+    (Lang::Por, POR),
+    (Lang::Ron, RON),
+    (Lang::Slk, SLK),
+    (Lang::Slv, SLV),
+    (Lang::Sna, SNA),
+    (Lang::Spa, SPA),
+    (Lang::Swe, SWE),
+    (Lang::Tgl, TGL),
+    (Lang::Tuk, TUK),
+    (Lang::Tur, TUR),
+    (Lang::Uzb, UZB),
+    (Lang::Vie, VIE),
+    (Lang::Zul, ZUL),
+];
 
-        _ => panic!("No alphabet for {}", lang),
-    };
-    alphabet.chars().collect()
-}
+/// Inverted map binding a character to a set of languages.
+pub static ALPHABET_LANG_MAP: Lazy<Vec<(char, Vec<Lang>)>> = Lazy::new(|| {
+    let mut map = HashMap::new();
+
+    for (lang, alphabet) in LATIN_ALPHABETS {
+        for c in alphabet.chars() {
+            let entry = map.entry(c).or_insert_with(Vec::new);
+            entry.push(*lang);
+        }
+    }
+
+    let mut char_lang = Vec::new();
+
+    for (ch, languages) in map {
+        char_lang.push((ch, languages));
+    }
+
+    char_lang
+});
 
 pub fn alphabet_calculate_scores(text: &LowercaseText, filter_list: &FilterList) -> RawOutcome {
+    let char_lang = &*ALPHABET_LANG_MAP;
+    let max_raw_score = text.chars().filter(|&ch| !is_stop_char(ch)).count();
+
     let mut raw_scores: Vec<(Lang, i32)> = Script::Latin
         .langs()
         .iter()
         .filter(|&&l| filter_list.is_allowed(l))
-        .map(|&l| (l, 0i32))
+        .map(|&l| (l, 0i32 - max_raw_score as i32))
         .collect();
 
-    let max_raw_score = text.chars().filter(|&ch| !is_stop_char(ch)).count();
+    for ch in text.chars() {
+        if is_stop_char(ch) {
+            continue;
+        }
 
-    for (lang, score) in &mut raw_scores {
-        let alphabet = get_lang_chars(*lang);
-
-        for ch in text.chars() {
-            if is_stop_char(ch) {
-                continue;
-            };
-            if alphabet.contains(&ch) {
-                *score += 1;
-            } else {
-                *score -= 1;
+        if let Some((_, languages)) = char_lang.iter().find(|(c, _)| *c == ch) {
+            for lang in languages {
+                if let Some((_, score)) = raw_scores.iter_mut().find(|(l, _)| l == lang) {
+                    *score += 2;
+                }
             }
         }
     }