--- a/src/alphabets/latin.rs
+++ b/src/alphabets/latin.rs
@@ -1,10 +1,12 @@
+use std::cmp::Reverse;
+use std::collections::HashMap;
+
+use once_cell::sync::Lazy;
+
 use super::RawOutcome;
 use crate::core::{FilterList, LowercaseText};
 use crate::utils::is_stop_char;
 use crate::{Lang, Script};
-
-use once_cell::sync::Lazy;
-use std::collections::HashMap;
 
 const AFR: &str = "abcdefghijklmnopqrstuvwxyzáèéêëíîïóôúû";
 const AKA: &str = "abdefghiklmnoprstuwyɔɛ";
@@ -112,8 +114,9 @@ pub fn alphabet_calculate_scores(text: &LowercaseText, filter_list: &FilterList)
     let (chars, langs) = &*ALPHABET_LANG_MAP;
 
     // score of each character.
+    let mut char_scores = vec![0; chars.len()];
     let mut max_raw_score = 0;
-    let mut scores: Vec<_> = chars.iter().map(|_| 0).collect();
+    // iterate over the text and scores characters.
     for ch in text.chars() {
         if is_stop_char(ch) {
             continue;
@@ -122,14 +125,17 @@ pub fn alphabet_calculate_scores(text: &LowercaseText, filter_list: &FilterList)
         max_raw_score += 1;
 
         if let Ok(position) = chars.binary_search(&ch) {
-            scores[position] += 2;
+            // add 2 and remove max_raw_score at the end,
+            // to keep the score interval of -max_raw_score..max_raw_score
+            char_scores[position] += 2;
         }
     }
 
-    let mut raw_scores: Vec<_> = (0..Lang::all().len()).into_iter().map(|_| 0).collect();
-
-    let mut common_score = 0;
-    for (position, char_score) in scores.into_iter().enumerate() {
+    // score of each lang.
+    let mut lang_scores = vec![0; Lang::all().len()];
+    let mut common_score: usize = 0;
+    // iterate over scored characters to compute language's scores.
+    for (position, char_score) in char_scores.into_iter().enumerate() {
         if char_score > 0 {
             let languages = &langs[position];
             // if current character is common to all Languages, increment a common score
@@ -137,28 +143,31 @@ pub fn alphabet_calculate_scores(text: &LowercaseText, filter_list: &FilterList)
             if languages.len() == LATIN_ALPHABETS.len() {
                 common_score += char_score;
             } else {
-                for lang in languages {
-                    raw_scores[*lang as usize] += char_score;
+                for &lang in languages {
+                    lang_scores[lang as usize] += char_score;
                 }
             }
         }
     }
 
+    // remap languages with theirs scores.
     let mut raw_scores: Vec<(Lang, usize)> = Script::Latin
         .langs()
         .iter()
         .filter(|&&l| filter_list.is_allowed(l))
-        .map(|&l| (l, raw_scores[l as usize]))
+        .map(|&l| {
+            let score = (lang_scores[l as usize] + common_score).saturating_sub(max_raw_score);
+            (l, score)
+        })
         .collect();
 
-    raw_scores.sort_unstable_by(|a, b| b.1.cmp(&a.1));
+    raw_scores.sort_unstable_by_key(|(_, score)| Reverse(*score));
 
     let mut normalized_scores = vec![];
 
-    for (lang, raw_score) in raw_scores.iter_mut() {
-        *raw_score = (*raw_score + common_score).saturating_sub(max_raw_score);
-        let normalized_score = *raw_score as f64 / max_raw_score as f64;
-        normalized_scores.push((*lang, normalized_score));
+    for &(lang, raw_score) in raw_scores.iter() {
+        let normalized_score = raw_score as f64 / max_raw_score as f64;
+        normalized_scores.push((lang, normalized_score));
     }
 
     RawOutcome {