@greyblake 2022-04-15
@ManyTheFish @Kerollmops 
Than you guys! This is very outstanding PR! I've learned something new today :)
Would you like to add further improvements? If you want, we can try to arrange a call, I can explain how trigrams work :)