@greyblake 2022-04-15
FYI: the optimization is released in 0.14.0.